fix(migrations): update like migration
