+++ backend/content-hub/alembic/versions/2025_04_22_1037-070ce4d0d6c3_add_like_table.py
@@ -1,0 +1,61 @@
+"""add like table
+
+Revision ID: 070ce4d0d6c3
+Revises: 2286c1510e2a
+Create Date: 2025-04-22 10:37:32.344472
+
+"""
+
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = "070ce4d0d6c3"
+down_revision: Union[str, None] = "2286c1510e2a"
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "like",
+        sa.Column("article_id", sa.Integer(), nullable=False),
+        sa.Column("user_id", sa.Integer(), nullable=False),
+        sa.Column(
+            "created_at", sa.DateTime(), server_default=sa.text("now()"), nullable=False
+        ),
+        sa.ForeignKeyConstraint(
+            ["article_id"], ["article.id"], name=op.f("fk_like_article_id_article")
+        ),
+        sa.ForeignKeyConstraint(
+            ["user_id"], ["user.id"], name=op.f("fk_like_user_id_user")
+        ),
+        sa.PrimaryKeyConstraint("article_id", "user_id", name=op.f("pk_like")),
+    )
+    op.alter_column(
+        "article",
+        "rating",
+        existing_type=sa.REAL(),
+        type_=sa.Float(precision=2),
+        existing_nullable=False,
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.alter_column(
+        "article",
+        "rating",
+        existing_type=sa.Float(precision=2),
+        type_=sa.REAL(),
+        existing_nullable=False,
+    )
+    op.drop_table("like")
+    # ### end Alembic commands ###
