+++ backend/content-hub/core/schemas/like.py
@@ -1,12 +1,15 @@
 from datetime import datetime, timezone
 from pydantic import BaseModel, ConfigDict, field_validator
 
+
 class LikeBase(BaseModel):
     article_id: int
     user_id: int
 
+
 class LikeCreate(LikeBase):
     model_config = ConfigDict(strict=True)
+
 
 class LikeOut(LikeBase):
     created_at: datetime
@@ -18,12 +21,9 @@ class LikeOut(LikeBase):
         return value
 
     class Config:
-        orm_mode = True
+        from_attributes = True
 
-like_data = {
-    "article_id": 1,
-    "user_id": 1,
-    "created_at": datetime.now(timezone.utc)
-}
+
+like_data = {"article_id": 1, "user_id": 1, "created_at": datetime.now(timezone.utc)}
 
 like = LikeOut(**like_data)
